🎨 영화 조회 api 추가

+++ src/main/java/com/example/drunkenmoviebackend/dto/reply/ReplyDto.java
@@ -10,7 +10,7 @@ import java.time.LocalDateTime;
 @AllArgsConstructor
 @Setter
 public class ReplyDto {
-    private Long replyId;
+    private Integer replyId;
     private Long userId;
     private String nickname;
     private String email;
